# docs/api-specification.yaml
---
openapi: 3.1.0
info:
  title: AI Resume Parser API
  version: 1.0.0
paths:
  "/health":
    get:
      summary: Health Check
      operationId: health_check_health_get
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
  "/api/v1/resumes/upload":
    post:
      tags:
        - Resumes
      summary: Upload Resume
      operationId: upload_resume_api_v1_resumes_upload_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              "$ref": "#/components/schemas/Body_upload_resume_api_v1_resumes_upload_post"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      security:
        - OAuth2PasswordBearer: []
  "/api/v1/resumes/{resume_id}":
    get:
      tags:
        - Resumes
      summary: Get Resume
      operationId: get_resume_api_v1_resumes__resume_id__get
      parameters:
        - name: resume_id
          in: path
          required: true
          schema:
            type: string
            title: Resume Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/api/v1/resumes/{id}/match":
    post:
      tags:
        - Resumes
      summary: Match Resume
      operationId: match_resume_api_v1_resumes__id__match_post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              title: Job Description
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/api/v1/resumes/analytics/{id}":
    get:
      tags:
        - Resumes
      summary: Analyze Resume
      operationId: analyze_resume_api_v1_resumes_analytics__id__get
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            title: Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/":
    get:
      summary: Root
      operationId: root__get
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
components:
  schemas:
    Body_upload_resume_api_v1_resumes_upload_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
        - file
      title: Body_upload_resume_api_v1_resumes_upload_post
    HTTPValidationError:
      properties:
        detail:
          items:
            "$ref": "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: token